Backend/requirements.txt
fastapi[standard]
uvicorn  
pydantic
langchain_core
langchain
langgraph
langchain-google-genai
requests
python-dotenv
speedtest-cli